requirements.txt
click==7.1.2
colorama==0.4.4
fastapi==0.63.0
gunicorn==20.1.0
h11==0.12.0
pydantic==1.8.1
python-dotenv==0.17.0
PyYAML==5.4.1
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.4
watchgod==0.7
websockets==8.1
httptools
google-cloud-storage
google-api-python-client
PyMySQL==0.9.3
google-auth
uvloop
google-api-python-client
google-auth-httplib2
google-auth-oauthlib